requirements.txt
absl-py @ file:///croot/absl-py_1714140470852/work
aiohappyeyeballs @ file:///croot/aiohappyeyeballs_1725434011349/work
aiohttp @ file:///croot/aiohttp_1725527756643/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
async-timeout @ file:///croot/async-timeout_1703096998144/work
attrs @ file:///croot/attrs_1729089401488/work
beautifulsoup4==4.12.3
blinker @ file:///croot/blinker_1696539051175/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools @ file:///croot/cachetools_1713977085444/work
cairocffi==1.7.1
CairoSVG==2.7.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///croot/click_1698129812380/work
coloredlogs==15.0.1
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cryptography @ file:///croot/cryptography_1724940562255/work
cssselect2==0.7.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
defusedxml==0.7.1
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1726613473834/work
flatbuffers==24.3.25
fonttools @ file:///croot/fonttools_1713551344105/work
frozenlist @ file:///croot/frozenlist_1698702560391/work
future @ file:///croot/future_1677599870788/work
gdown==5.2.0
google-auth @ file:///croot/google-auth_1715111149956/work
google-auth-oauthlib @ file:///opt/conda/conda-bld/google-auth-oauthlib_1660687784486/work
grpcio @ file:///croot/grpc-suite_1701982571719/work
h5py==3.11.0
humanfriendly==10.0
idna @ file:///croot/idna_1714398848350/work
imageio==2.35.1
imgviz==1.7.5
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib_resources @ file:///croot/importlib_resources-suite_1720641103994/work
imutils==0.5.4
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
joblib==1.4.2
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lazy_loader==0.4
lxml @ file:///croot/lxml_1722882187815/work
Markdown @ file:///croot/markdown_1671541909495/work
markdown-it-py==3.0.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1706899923320/work
matplotlib==3.7.5
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
multidict @ file:///croot/multidict_1701096859099/work
natsort==8.4.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1680692919326/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
oauthlib @ file:///croot/oauthlib_1679489621486/work
onnx==1.17.0
onnx-simplifier==0.4.36
onnxruntime==1.19.2
opencv-python==4.1.2.30
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.0.3
pillow==10.4.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
ply==3.11
pooch @ file:///croot/pooch_1695850093751/work
protobuf==3.20.3
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycairo==1.26.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.18.0
PyGObject==3.48.2
PyJWT @ file:///croot/pyjwt_1715094751094/work
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyOpenSSL @ file:///croot/pyopenssl_1723651579422/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
pypotrace==0.3
PyQt5==5.15.11
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.15.0
PySide6==6.6.3.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.0.1
pytz==2024.2
PyWavefront==1.3.3
PyWavelets==1.4.1
PyYAML @ file:///croot/pyyaml_1728657952215/work
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
QtPy==2.4.1
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib @ file:///croot/requests-oauthlib_1720615017878/work
rich==13.9.3
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shiboken6==6.6.3.1
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.6
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1728484454203/work
tensorboard @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorboard_1682445826165/work/tensorboard-2.12.1-py3-none-any.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorboard_data_server @ file:///croot/tensorboard-data-server_1724172556187/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=7da966c3e116be7b4d44cd25a804e06f45fa6d4a1f03819052dd9cb3482b141a
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.3.0
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.66.5
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1727769808118/work
vtk==9.3.1
webencodings==0.5.1
Werkzeug @ file:///croot/werkzeug_1716533313107/work
xlrd @ file:///croot/xlrd_1685030938141/work
yarl @ file:///croot/yarl_1725976495189/work
zipp @ file:///croot/zipp_1729012354496/work
